fixes in v680:
1) changed frame delimiter from '##' to '0##' because of 'dropped'
messages caused by the null value after imei that turned in an ## after
the imei. I am not sure if there is always a 0 before the final ## but
in hours of testing it was always true for my v680 clone.

2) trimmed sentence in v680 protocol decoder to fix 'dropped' messages
caused by line feeds that my v680 clone was receiving before the message
itself.