Added support for Alarm popup & positions history cleaner
1. added support for Alarm and Alarm Type in popup and also in history
records. can be muted with a toggle button in the UI
2. added a timer to clear positions history once a day. the default is
to save positions history for 7 days if not defined in the configuration
file.
3. prevent the lock of the UI files (js, html, css, etc..) for wondows
developers in debug mode. It's easier to do changes to the UI without
stopping the app each time.
4. tools: added support in test-generator.py to simulate Alert, also
added minify.bat file to compile new js files for the UI using sencha
SDK for windows developers