T580W: add battery level to all events
It is possible to use "processing.copyAttributes" configuration to have
battery level in all events, but this will not work if battery events will
be filtered somehow and dropped from DB, so it will be no event in DB with
latest battery level. So, ensure all our events have latest battery level.