org.traccar.DistanceHandler: add parameter coordinates.minErrorStatic
This commit adds the following configuration parameter:

coordinates.minErrorStatic
        Distance in meters. Value of `coordinates.minErrorStatic` that
        applies when the speed *as reported by the device* is equal to
        zero (or, more exactly, less then 0.01 knots). By default this
        value is equal to `coordinates.minErrorStatic`.

Using this parameter works very well for filtering out erroneous position changes
when the device is static. As we all know, GPS is imprecise and using it
introduces position errors. When the device is static, every position
recorded by GPS is slightly different and position history - when displayed
on a map - looks like a distinctive "star".
In this situation, most device position updates feature a speed of (exactly) 0.0 knots.
By replacing all positions with a speed of 0.0 knots with last
position, we can make the device appear almost "static" on the map.

It is safe to use much higher values of `coordinates.minErrorStatic` than
values of `coordinates.minError`.

This filtering mechanism works very well in practice. It is not
necessary to make the threshold speed user-configurable.
It worked very well even if the comparison

    speed == 0.0

was used.

This work was funded by Partner Security (www.partnersecurity.pl) but
copyright remains with me (Mateusz Jończyk). Additionally, I have
obtained permission to release my work as open source.