fix watchframedecoder for TK906 (invalid lenght) device error
This fixes device issues found with TK906 bike trackers, which tend to
transmit wrong a frame length for some frames very often. 

The fixed Watch protocol frame decoder now not only evaluates the frame
length transmitted from the device itself but also checks the no of
frame bytes till the next "frame end marker" (']') and uses whatever is
correct.

Additionally the frame start detection has been improved to better sync
to valid frames according to the watch protocol spec.

usecase 1) ignore chars which don't belong to a frame (not between '['
and ']'):

12345[SG*12345678*0004*LK,,]abcde\n..[SG*12345678*0004*LK,,]222

in this example '12345', 'abcde\n..' and '222' stream bytes will be
ignored

usecase 2) partial transmitted frames

[SG*12345678[SG*12345678*0002*LK]

in this example, the leading bytes '[SG*12345678' will be ignored


The fix does not affact devices which work according to the protocol
spec, but improves handling of devices like the TK906 which fail to
fulfill the protcol spec and otherwise could not be used with traccar.