fix and improve watch - voice msg cmd handling
+-----------------------------------------------------------------+
fix: 
====

The cmd was not sent before when triggered from the web frontend if
"real" amr data (taken from a file generated from a smartwatch) was
entered to the "Key_Data" field within the webform, for some reasons:

the call (WatchProtocolEnecoder.java::191): 
cmdtoSend = formatCommand(channel, command, "TK," +
getBinaryData(command));

directly concats the binay data to the cmd content, which causes (with
real data) a handling problem within the class
StringProtocolEncoder::formatCommand
The binary blob also can contain "\" and "$" which cause problems 
when a string replacement methode is called.

According to java doc 'quoteReplacement' should be used in that cases...

Also changed the direct content concating to using the "key -
replacment conecept" all other cmds use.

Also changed the return type of the encode methode to "bytebuf" because
strings cause problems on all non-ASCII data - e.g. as the binary blob
of an amr file.

For the same reason, also the unit test of the watchencoder has been
changed from string to binary assert values and the voice msg test now
covers non-ascii char + char which need to be escaped.


This changes don't affact usual text cmds (which don't contain binary
data) - according to my tests but make the VOICE MSG cmd usable.
+-----------------------------------------------------------------+
Improvement: 
============
Introduced a new cmd type (binary blob) for e.g. voice data.

Also added a way to specify a filename instead of forcing the user to
copy & paste unhandy audio data as hex encoded stream to the web form,
because it should be easier to reference to a file (and read the binary
blob from this file) than to extract this data manually.

To reference an amr - audio file, one needs to enter
"file://[rel.path]/[filename.amr]" within the "KEY_DATA" field in the
webform.
The path of the file is relative to the one which is configured as 
"media.path" 

e.g. if media.path is set to "/mypath" and the audio file is
stored in /mypath/foo/bar.amr then "file://foo/bar.amr" entered
as KEY_DATA will trigger a file read.
+-----------------------------------------------------------------+