tlt2h: Fix UDP
Before this change, the tlt2h decoder did not correctly detect the end
of the datagram, leading to an additional empty position returned in
positions. Since wifi networks are not cached, this resulted in networks
being never showed in the `details` view of a device. This was initially
reported in https://github.com/traccar/traccar/pull/5344#issuecomment-2266344264